+++ lib/chatview.dart
@@ -35,6 +35,6 @@ export 'package:audio_waveforms/audio_waveforms.dart'
         AndroidEncoder,
         IosEncoder,
         AndroidOutputFormat;
-export 'src/models/receipts_widget_config.dart';
+export 'src/models/config_models/receipts_widget_config.dart';
 export 'src/extensions/extensions.dart' show MessageTypes;
 export 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
