+++ lib/src/models/config_models/chat_view_states_configuration.dart
