+++ lib/src/models/config_models/emoji_message_configuration.dart
