+++ lib/src/models/config_models/feature_active_config.dart
@@ -53,4 +53,5 @@ class FeatureActiveConfig {
 
   /// Used for enable/disable other users name.
   final bool enableOtherUserName;
+
 }
