+++ lib/src/models/config_models/image_message_configuration.dart
@@ -21,7 +21,7 @@
  */
 import 'package:flutter/material.dart';
 
-import '../values/typedefs.dart';
+import '../../values/typedefs.dart';
 
 class ImageMessageConfiguration {
   /// Provides configuration of share button while image message is appeared.
