+++ lib/src/models/config_models/link_preview_configuration.dart
