+++ lib/src/models/config_models/profile_circle_configuration.dart
@@ -19,28 +19,37 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
+import 'package:flutter/material.dart';
 
-import '../utils/constants/constants.dart';
-import '../values/enumeration.dart';
-import '../values/typedefs.dart';
+import '../../../chatview.dart';
+import '../../utils/constants/constants.dart';
 
-class ChatUser {
-  /// Provides id of user.
-  final String id;
+class ProfileCircleConfiguration {
+  /// Used to give padding to profile circle.
+  final EdgeInsetsGeometry? padding;
 
-  /// Provides name of user.
-  final String name;
-
-  /// Provides profile picture as network URL or asset of user.
+  /// Provides image url as network or asset of user.
   /// Or
-  /// Provides profile picture's data in base64 string.
-  final String? profilePhoto;
+  /// Provides image data of user in base64
+  final String? profileImageUrl;
 
-  /// Field to set default image if network url for profile image not provided
-  final String defaultAvatarImage;
+  /// Used for give bottom padding to profile circle
+  final double? bottomPadding;
+
+  /// Used for give circle radius to profile circle
+  final double? circleRadius;
+
+  /// Provides callback when user tap on profile circle.
+  final void Function(ChatUser)? onAvatarTap;
+
+  /// Provides callback when user long press on profile circle.
+  final void Function(ChatUser)? onAvatarLongPress;
 
   /// Field to define image type [network, asset or base64]
   final ImageType imageType;
+
+  /// Field to set default avatar image if profile image link not provided
+  final String defaultAvatarImage;
 
   /// Error builder to build error widget for asset image
   final AssetImageErrorBuilder? assetImageErrorBuilder;
@@ -52,49 +61,17 @@ class ChatUser {
   final NetworkImageProgressIndicatorBuilder?
       networkImageProgressIndicatorBuilder;
 
-  ChatUser({
-    required this.id,
-    required this.name,
-    this.profilePhoto,
-    this.defaultAvatarImage = profileImage,
+  const ProfileCircleConfiguration({
+    this.onAvatarTap,
+    this.padding,
+    this.profileImageUrl,
+    this.bottomPadding,
+    this.circleRadius,
+    this.onAvatarLongPress,
     this.imageType = ImageType.network,
-    this.assetImageErrorBuilder,
+    this.defaultAvatarImage = profileImage,
     this.networkImageErrorBuilder,
+    this.assetImageErrorBuilder,
     this.networkImageProgressIndicatorBuilder,
   });
-
-  factory ChatUser.fromJson(Map<String, dynamic> json) => ChatUser(
-        id: json["id"],
-        name: json["name"],
-        profilePhoto: json["profilePhoto"],
-        imageType: ImageType.tryParse(json['imageType']?.toString()) ??
-            ImageType.network,
-        defaultAvatarImage: json["defaultAvatarImage"],
-      );
-
-  Map<String, dynamic> toJson() => {
-        'id': id,
-        'name': name,
-        'profilePhoto': profilePhoto,
-        'imageType': imageType.name,
-        'defaultAvatarImage': defaultAvatarImage,
-      };
-
-  ChatUser copyWith({
-    String? id,
-    String? name,
-    String? profilePhoto,
-    ImageType? imageType,
-    String? defaultAvatarImage,
-    bool forceNullValue = false,
-  }) {
-    return ChatUser(
-      id: id ?? this.id,
-      name: name ?? this.name,
-      imageType: imageType ?? this.imageType,
-      profilePhoto:
-          forceNullValue ? profilePhoto : profilePhoto ?? this.profilePhoto,
-      defaultAvatarImage: defaultAvatarImage ?? this.defaultAvatarImage,
-    );
-  }
 }
