+++ lib/src/models/config_models/receipts_widget_config.dart
@@ -21,8 +21,8 @@
  */
 
 import 'package:flutter/cupertino.dart';
-import '../../chatview.dart';
-import '../utils/constants/constants.dart';
+import '../../../chatview.dart';
+import '../../utils/constants/constants.dart';
 
 class ReceiptsWidgetConfig {
   /// The builder that builds widget that right next to the senders message bubble.
