+++ lib/src/models/config_models/replied_msg_auto_scroll_config.dart
