+++ lib/src/models/config_models/suggestion_list_config.dart
@@ -1,4 +1,4 @@
-import '../values/enumeration.dart';
+import '../../values/enumeration.dart';
 import 'package:flutter/cupertino.dart';
 
 class SuggestionListConfig {
