+++ lib/src/models/config_models/swipe_to_reply_configuration.dart
