+++ lib/src/models/config_models/type_indicator_configuration.dart
