+++ lib/src/models/models.dart
@@ -20,28 +20,28 @@
  * SOFTWARE.
  */
 export 'chat_bubble.dart';
-export 'image_message.dart';
-export 'message.dart';
-export 'message_reaction_configuration.dart';
-export 'reply_message.dart';
-export 'profile_circle.dart';
-export 'chat_bubble_configuration.dart';
-export 'replied_message_configuration.dart';
-export 'swipe_to_reply_configuration.dart';
-export 'reply_popup_configuration.dart';
-export 'reaction_popup_configuration.dart';
-export 'message_list_configuration.dart';
-export 'emoji_message_configuration.dart';
-export 'message_configuration.dart';
-export 'send_message_configuration.dart';
-export 'chat_user.dart';
-export 'link_preview_configuration.dart';
-export 'type_indicator_configuration.dart';
-export 'chat_view_states_configuration.dart';
-export 'reaction.dart';
-export 'replied_msg_auto_scroll_config.dart';
-export 'feature_active_config.dart';
-export 'suggestion_item_data.dart';
-export 'reply_suggestions_config.dart';
-export 'suggestion_list_config.dart';
-export 'suggestion_item_config.dart';
+export 'config_models/image_message_configuration.dart';
+export 'data_models/message.dart';
+export 'config_models/message_reaction_configuration.dart';
+export 'data_models/reply_message.dart';
+export 'config_models/profile_circle_configuration.dart';
+export 'config_models/chat_bubble_configuration.dart';
+export 'config_models/replied_message_configuration.dart';
+export 'config_models/swipe_to_reply_configuration.dart';
+export 'config_models/reply_popup_configuration.dart';
+export 'config_models/reaction_popup_configuration.dart';
+export 'config_models/message_list_configuration.dart';
+export 'config_models/emoji_message_configuration.dart';
+export 'config_models/message_configuration.dart';
+export 'config_models/send_message_configuration.dart';
+export 'data_models/chat_user.dart';
+export 'config_models/link_preview_configuration.dart';
+export 'config_models/type_indicator_configuration.dart';
+export 'config_models/chat_view_states_configuration.dart';
+export 'data_models/reaction.dart';
+export 'config_models/replied_msg_auto_scroll_config.dart';
+export 'config_models/feature_active_config.dart';
+export 'data_models/suggestion_item_data.dart';
+export 'config_models/reply_suggestions_config.dart';
+export 'config_models/suggestion_list_config.dart';
+export 'config_models/suggestion_item_config.dart';
