+++ lib/src/widgets/link_preview.dart
@@ -21,7 +21,7 @@
  */
 import 'package:any_link_preview/any_link_preview.dart';
 import 'package:chatview/src/extensions/extensions.dart';
-import 'package:chatview/src/models/link_preview_configuration.dart';
+import 'package:chatview/src/models/config_models/link_preview_configuration.dart';
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
