+++ lib/src/widgets/share_icon.dart
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 import 'package:flutter/material.dart';
-import 'package:chatview/src/models/image_message.dart';
+import 'package:chatview/src/models/config_models/image_message_configuration.dart';
 
 class ShareIcon extends StatelessWidget {
   const ShareIcon({
