+++ lib/src/widgets/suggestions/suggestions_config_inherited_widget.dart
@@ -1,4 +1,4 @@
-import 'package:chatview/src/models/reply_suggestions_config.dart';
+import 'package:chatview/src/models/config_models/reply_suggestions_config.dart';
 import 'package:flutter/material.dart';
 import 'package:chatview/chatview.dart';
